refactor(app): move UpdateRobot
