+++ app/src/App/OnDeviceDisplayApp.tsx
@@ -38,8 +38,8 @@ import { QuickTransferDashboard } from '../pages/ODD/QuickTransferDashboard'
 import { QuickTransferDetails } from '../pages/ODD/QuickTransferDetails'
 import { RunningProtocol } from '../pages/RunningProtocol'
 import { RunSummary } from '../pages/RunSummary'
-import { UpdateRobot } from '../pages/UpdateRobot/UpdateRobot'
-import { UpdateRobotDuringOnboarding } from '../pages/UpdateRobot/UpdateRobotDuringOnboarding'
+import { UpdateRobot } from '../pages/ODD/UpdateRobot/UpdateRobot'
+import { UpdateRobotDuringOnboarding } from '../pages/ODD/UpdateRobot/UpdateRobotDuringOnboarding'
 import { InstrumentsDashboard } from '../pages/ODD/InstrumentsDashboard'
 import { InstrumentDetail } from '../pages/ODD/InstrumentDetail'
 import { Welcome } from '../pages/Welcome'
