+++ app/src/pages/ODD/UpdateRobot/UpdateRobot.tsx
@@ -5,22 +5,22 @@ import { useTranslation } from 'react-i18next'
 
 import { Flex, SPACING, DIRECTION_ROW } from '@opentrons/components'
 
-import { getLocalRobot } from '../../redux/discovery'
+import { getLocalRobot } from '../../../redux/discovery'
 import {
   getRobotUpdateAvailable,
   clearRobotUpdateSession,
-} from '../../redux/robot-update'
-import { useDispatchStartRobotUpdate } from '../../redux/robot-update/hooks'
-import { UNREACHABLE } from '../../redux/discovery/constants'
+} from '../../../redux/robot-update'
+import { useDispatchStartRobotUpdate } from '../../../redux/robot-update/hooks'
+import { UNREACHABLE } from '../../../redux/discovery/constants'
 
-import { MediumButton } from '../../atoms/buttons'
+import { MediumButton } from '../../../atoms/buttons'
 import {
   UpdateRobotSoftware,
   NoUpdateFound,
   ErrorUpdateSoftware,
-} from '../../organisms/UpdateRobotSoftware'
+} from '../../../organisms/UpdateRobotSoftware'
 
-import type { State, Dispatch } from '../../redux/types'
+import type { State, Dispatch } from '../../../redux/types'
 
 export function UpdateRobot(): JSX.Element {
   const navigate = useNavigate()
