+++ app/src/pages/ODD/UpdateRobot/__tests__/UpdateRobotDuringOnboarding.test.tsx
@@ -2,19 +2,19 @@ import * as React from 'react'
 import { vi, it, describe, expect, beforeEach, afterEach } from 'vitest'
 import { MemoryRouter } from 'react-router-dom'
 import { act, screen } from '@testing-library/react'
-import { renderWithProviders } from '../../../__testing-utils__'
-import { i18n } from '../../../i18n'
+import { renderWithProviders } from '../../../../__testing-utils__'
+import { i18n } from '../../../../i18n'
 
-import * as RobotUpdate from '../../../redux/robot-update'
-import type { RobotUpdateSession } from '../../../redux/robot-update/types'
-import { getLocalRobot } from '../../../redux/discovery'
+import * as RobotUpdate from '../../../../redux/robot-update'
+import type { RobotUpdateSession } from '../../../../redux/robot-update/types'
+import { getLocalRobot } from '../../../../redux/discovery'
 
 import { UpdateRobotDuringOnboarding } from '../UpdateRobotDuringOnboarding'
 
-import type { State } from '../../../redux/types'
+import type { State } from '../../../../redux/types'
 
-vi.mock('../../../redux/discovery')
-vi.mock('../../../redux/robot-update')
+vi.mock('../../../../redux/discovery')
+vi.mock('../../../../redux/robot-update')
 
 const MOCK_STATE: State = {
   discovery: {
